feat: move volume commands to media extension
